requirements.txt
anthropic>=0.39.0
requests>=2.31.0
PyYAML>=6.0.1
PyGithub>=2.1.1
Jinja2>=3.1.2
python-dateutil>=2.8.2
